FINALLY HAVE A SHOPPING CART - WOOOHOOOOOOO

+++ HayleesThreads/Models/Product.cs
@@ -15,7 +15,7 @@ namespace HayleesThreads.Models
 
     public string ProductName { get; set; }
   
-    public float ProductPrice { get; set; }
+    public decimal ProductPrice { get; set; }
 
     public string ProductImage { get; set; }
